requirements.txt
flask
openai
twilio
python-dotenv
faiss-cpu

